+++ src/by/tsvrko/manics/dao/interfaces/db/ChatDAO.java
@@ -13,7 +13,7 @@ public interface ChatDAO {
 
      boolean addChat(ChatInfo chatInfo, String token);
 
-     Chat getChatById(int chatId);
+     Chat getChatById(long chatId);
 
      boolean deleteChat(Chat chat);
